added cart functionality and functions moved to helpers
