+++ src/App.tsx
@@ -27,60 +27,6 @@ export const DataContext = createContext<DataContextType>({
   setCartStorage: () => { },
 });
 
-export function handleAddToCart(item: Product,
-  setCartStorage: React.Dispatch<React.SetStateAction<Product[]>>): void {
-  try {
-    const currentCart: Product[]
-      = JSON.parse(localStorage.getItem('cart') || '[]');
-
-    const isItemInCart = currentCart.some(product => product.id === item.id);
-
-    if (isItemInCart) {
-      const updatedCart = currentCart.filter(product => product.id !== item.id);
-
-      localStorage.setItem('cart', JSON.stringify(updatedCart));
-      setCartStorage(updatedCart);
-
-      return;
-    }
-
-    const updatedCart = [...currentCart, item];
-
-    localStorage.setItem('cart', JSON.stringify(updatedCart));
-    setCartStorage(updatedCart);
-  } catch (error) {
-    throw new Error('Error updating cart data');
-  }
-}
-
-export function handleAddToFavorites(item: Product,
-  setFavoriteStorage: React.Dispatch<React.SetStateAction<Product[]>>): void {
-  try {
-    const currentFavorites: Product[]
-      = JSON.parse(localStorage.getItem('favorites') || '[]');
-
-    const isItemInFavorites
-      = currentFavorites.some(product => product.id === item.id);
-
-    if (isItemInFavorites) {
-      const updatedFavorites
-        = currentFavorites.filter(product => product.id !== item.id);
-
-      localStorage.setItem('favorites', JSON.stringify(updatedFavorites));
-      setFavoriteStorage(updatedFavorites);
-
-      return;
-    }
-
-    const updatedFavorites = [...currentFavorites, item];
-
-    localStorage.setItem('favorites', JSON.stringify(updatedFavorites));
-    setFavoriteStorage(updatedFavorites);
-  } catch (error) {
-    throw new Error('Error updating favorites data');
-  }
-}
-
 export const App = () => {
   const [productList, setProductList] = useState<Product[]>([]);
   const [isLoading, setIsLoading] = useState(true);
