+++ src/components/AddToCart/AddToCart.tsx
@@ -4,7 +4,8 @@ import './AddToCart.scss';
 import like from '../../icons/like.svg';
 import likeActive from '../../icons/like-active.svg';
 import { Product } from '../../types/product';
-import { DataContext, handleAddToCart, handleAddToFavorites } from '../../App';
+import { DataContext } from '../../App';
+import { handleAddToCart, handleAddToFavorites } from '../../utils/helpers';
 
 type Props = {
   product: Product,
