+++ src/types/product.ts
@@ -11,4 +11,5 @@ export interface Product {
   ram: string;
   year: number;
   image: string;
+  quantity?: number;
 }
